+++ OneSignalExample/Assets/OneSignal/CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
-- Updated included Android SDK to [4.7.4](https://github.com/OneSignal/OneSignal-Android-SDK/releases/tag/4.7.4)
+- Added support for OneSignal Android functionality `promptForPushNotifications`
+- Updated included Android SDK to [4.8.0](https://github.com/OneSignal/OneSignal-Android-SDK/releases/tag/4.8.0)
 - Updated included iOS SDK to [3.11.2](https://github.com/OneSignal/OneSignal-iOS-SDK/releases/tag/3.11.2)
 - Added support for OneSignal Android `setLanguage` callbacks